All algorithms (sans round robin until Tyler updates the repo) have been moved into their own class files.
