+++ src/schedulingAlgorithms/HighestPriorityFirst.java
@@ -1,0 +1,23 @@
+package schedulingAlgorithms;
+
+public class HighestPriorityFirst 
+{
+	public HighestPriorityFirst()
+	{
+		
+	}
+	
+    /**
+     * Highest Priority First (non-preemptive)
+     */
+    private void preemptive() {
+        //TODO 
+    }
+
+    /**
+     * Higest Priority First (preemptive)
+     */
+    private void nonPreemptive() {
+        //TODO
+    }
+}
