+++ src/schedulingAlgorithms/ShortestRemainingTime.java
@@ -1,0 +1,8 @@
+package schedulingAlgorithms;
+
+public class ShortestRemainingTime {
+    //Shortest remaining time (preemptive)
+    private void ShortestRemainingTime() {
+        //TODO
+    }
+}
